Switch to uint from int as quantities cannt be negative (#489)
Switch to uint from int as quantities cannt be negative (#489)

+++ contracts/LPToken.sol
@@ -17,22 +17,22 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
 
     function lpTokenForUser(
         address _user
-    ) external view override returns (int256) {
-        return int256(lpToken.balanceOf(_user));
+    ) external view override returns (uint256) {
+        return lpToken.balanceOf(_user);
     }
 
     function getLpTokenAddress() external view override returns (address) {
         return address(lpToken);
     }
 
-    function getAllLPTokenCount() external view override returns (int256) {
-        return int256(lpToken.totalSupply());
+    function getAllLPTokenCount() external view override returns (uint256) {
+        return lpToken.totalSupply();
     }
 
     function lpTokenCountForGivenTokensQty(
-        int256 tokenAQuantity,
-        int256 tokenBQuantity
-    ) external pure override returns (int256) {
+        uint256 tokenAQuantity,
+        uint256 tokenBQuantity
+    ) external pure override returns (uint256) {
         return sqrt(tokenAQuantity * tokenBQuantity);
     }
 
@@ -60,22 +60,22 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
     }
 
     function allotLPTokenFor(
-        int256 amountA,
-        int256 amountB,
+        uint256 amountA,
+        uint256 amountB,
         address _toUser
     ) external override onlyOwner returns (int256 responseCode) {
         require(
             (amountA > 0 && amountB > 0),
             "Please provide positive token counts"
         );
-        int256 mintingAmount = this.lpTokenCountForGivenTokensQty(
+        uint256 mintingAmount = this.lpTokenCountForGivenTokensQty(
             amountA,
             amountB
         );
         (responseCode, ) = super.mintToken(
             tokenService,
             address(lpToken),
-            mintingAmount
+            int256(mintingAmount)
         );
         require(
             responseCode == HederaResponseCodes.SUCCESS,
@@ -85,7 +85,7 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
             address(lpToken),
             address(this),
             _toUser,
-            mintingAmount
+            int256(mintingAmount)
         );
         require(
             responseCode == HederaResponseCodes.SUCCESS,
@@ -95,7 +95,7 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
     }
 
     function removeLPTokenFor(
-        int256 lpAmount,
+        uint256 lpAmount,
         address fromUser
     ) external override onlyOwner returns (int256 responseCode) {
         require((lpAmount > 0), "Please provide token counts");
@@ -108,7 +108,7 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
             address(lpToken),
             fromUser,
             address(this),
-            lpAmount
+            int256(lpAmount)
         );
 
         require(
@@ -120,7 +120,7 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
         (responseCode, ) = super.burnToken(
             tokenService,
             address(lpToken),
-            lpAmount
+            int256(lpAmount)
         );
         require(
             responseCode == HederaResponseCodes.SUCCESS,
@@ -129,8 +129,8 @@ contract LPToken is ILPToken, OwnableUpgradeable, TokenOperations {
         return HederaResponseCodes.SUCCESS;
     }
 
-    function sqrt(int256 value) private pure returns (int256 output) {
-        int256 modifiedValue = (value + 1) / 2;
+    function sqrt(uint256 value) private pure returns (uint256 output) {
+        uint256 modifiedValue = (value + 1) / 2;
         output = value;
         while (modifiedValue < output) {
             output = modifiedValue;
